fix: move docs
