+++ packages/hydrogen/src/foundation/useQuery/docs/1-arguments.md
