+++ packages/hydrogen/src/foundation/useShopQuery/README.md
