+++ packages/hydrogen/src/foundation/useShopQuery/docs/1-arguments.md
