+++ packages/hydrogen/src/foundation/useShopQuery/docs/2-return-value.md
