+++ packages/hydrogen/src/foundation/useShopQuery/examples/use-shop-query.example.tsx
