+++ packages/hydrogen/src/hooks/index.ts
@@ -1,6 +1,6 @@
 export {useProductOptions} from './useProductOptions';
 export * from './useProductOptions/types';
-export {useQuery, QueryProvider} from './useQuery';
+export {useQuery, QueryProvider} from '../foundation/useQuery';
 
 export {useMoney} from './useMoney';
 export {useMeasurement} from './useMeasurement';
